requirements.txt
torch>=2.6.0
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0
yfinance>=0.2.31
pyyaml>=6.0
matplotlib>=3.7.2
tqdm>=4.65.0